requirements.txt
boto3==1.35.81
botocore==1.35.81
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==1.26.20
ollama==0.4.4
